requirements.txt
asgiref==3.11.0
dataclasses==0.6
Django==6.0
djongo==1.2.31
dnspython==2.8.0
pymongo==4.15.5
sqlparse==0.5.5
tzdata==2025.3
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8